+++ src/include/config/constants.py
@@ -31,7 +31,7 @@ from pathlib import Path
 from include.config.version import Version
 
 CORE_VERSION = Version("0.5.0.260820_alpha")
-PROTOCOL_VERSION = 23
+PROTOCOL_VERSION = 24
 
 ROOT_ABSPATH = Path(__file__).resolve().parents[2]