+++ src/include/domains/identity/commands/users.py
@@ -58,7 +58,7 @@ def create_user(**kwargs) -> None:
                 permission=i["permission"],
                 granted=i.get("granted", True),
                 start_time=i.get("start_time", time.time()),
-                end_time=i.get("end_time", None),
+                end_time=i.get("end_time"),
             )
             user.rights.append(permission)